+++ Assets/Scripts/UI/DialogueSystem.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 5d70833988199c949bc9c1bd24aae882
+timeCreated: 1518501070
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
